+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## master (unreleased)
+* Updates default ruby version of GitHub Actions install generator. ([@jamesglover][])
 * Simplify generated GitHub Actions install ([@jamesglover][])
 
 ## 0.5.0 (December 20th, 2020)